fuzz tests debug fix (#351)
* new tests and fix

* add tests

* add new test

* fix

* remove error not used

* disable wrong tests

* fix

* remove extra check

* fix tests

* don't stop after fuzzer stop

* try fix

* fix assurance validation

+++ Blockchain/Sources/Blockchain/VMInvocations/Invocations/AccumulateInvocation.swift
@@ -31,7 +31,7 @@ public func accumulate(
         return .init(state: state, transfers: [], commitment: nil, gasUsed: Gas(0), provide: [])
     }
 
-    let initialIndex = try Blake2b256.hash(JamEncoder.encode(serviceIndex, state.entropy, timeslot)).data.decode(UInt32.self)
+    let initialIndex = try Blake2b256.hash(JamEncoder.encode(UInt(serviceIndex), state.entropy, UInt(timeslot))).data.decode(UInt32.self)
     let nextAccountIndex = try await AccumulateContext.check(
         i: (initialIndex % serviceIndexModValue) + 256,
         accounts: state.accounts.toRef()
